参与者与系统用户整合

+++ openhis-server/openhis-application/src/main/java/com/openhis/web/basedatamanage/appservice/impl/PractitionerAppServiceImpl.java
@@ -224,6 +224,9 @@ public class PractitionerAppServiceImpl implements IPractitionerAppService {
      */
     @Override
     public R<?> delUserPractitioner(Long userId) {
+        if (1L == userId) {
+            return R.fail(null, "admin不允许删除");
+        }
         iBizUserService.remove(new LambdaQueryWrapper<BizUser>().eq(BizUser::getUserId, userId));
         practitionerAppAppMapper.delUserRole(userId);
         Practitioner one =
